Move paintable image moc class from cpp to .h
